重构http初始化方式,使扩展性更强
