+++ lib/ui/page/change_log_page.dart
@@ -1,8 +1,11 @@
+import 'dart:io';
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_markdown/flutter_markdown.dart';
 import 'package:fun_android/generated/i18n.dart';
+import 'package:fun_android/view_model/app_model.dart';
 
 class ChangeLogPage extends StatelessWidget {
   Widget build(BuildContext context) {
@@ -22,10 +25,20 @@ class ChangeLogPage extends StatelessWidget {
             bottom: 8,
             child: CupertinoButton(
               color: Theme.of(context).accentColor,
-              onPressed: () {
-                Navigator.pop(context);
+              onPressed: () async {
+                if (Platform.isAndroid) {
+                  /// TODO 1.增加Loading状态
+                  String url = await CheckUpdateModel().checkUpdate();
+                  if (url?.isNotEmpty ?? false) {
+                    /// 通过flutter_downloader 1.2.1下载
+                  }
+                } else {
+                  Navigator.pop(context);
+                }
               },
-              child: Text(S.of(context).close),
+              child: Text(Platform.isAndroid
+                  ? S.of(context).checkUpdate
+                  : S.of(context).close),
             ),
           )
         ]),
