+++ lib/view_model/app_model.dart
@@ -1,4 +1,8 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
+import 'package:fun_android/service/app_repository.dart';
+import 'package:fun_android/utils/platform_utils.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 const kAppFirstEntry = 'kAppFirstEntry';
@@ -12,4 +16,13 @@ class AppModel with ChangeNotifier {
     isFirst = sharedPreferences.getBool(kAppFirstEntry);
     notifyListeners();
   }
+}
+
+class CheckUpdateModel with ChangeNotifier {
+
+  checkUpdate() async{
+    var appVersion = await PlatformUtils.getAppVersion();
+    return AppRepository.checkUpdate(Platform.operatingSystem, appVersion);
+  }
+
 }
